--- a/.github/scripts/test-config.js
+++ b/.github/scripts/test-config.js
@@ -81,7 +81,7 @@ const MOCK_RESPONSES = {
 };
 
 // Export configuration
-module.exports = {
+export {
   TEST_CONFIG,
   setupTestEnvironment,
   MOCK_RESPONSES