… .github/scripts/activity-analyzer.test.js
… .github/scripts/claude-enhancer.test.js
… .github/scripts/comprehensive-issue-grooming-august-3-2025.md
… .github/scripts/cv-generator.test.js
… .github/scripts/cv-performance-optimizer.js
… .github/scripts/data/auth-performance-optimization.json
… .github/scripts/data/career-intelligence/job-market-analysis-2025-08-03.json
… .github/scripts/data/career-pathways/career-pathway-analysis-2025-08-03.json
… .github/scripts/data/cv-performance-optimization.json
… .github/scripts/data/gemini-playwright-audit.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754196405561.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754202069928.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754204905929.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754205858031.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206137630.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206161923.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206184903.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206467705.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206485089.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206493665.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206502613.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206530147.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206538820.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206547544.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754206556356.json
… .github/scripts/data/networking-intelligence/brand-optimization-1754207034368.json
… .github/scripts/data/networking-intelligence/test-validation-1754196405568.json
… .github/scripts/data/networking-intelligence/test-validation-1754202069933.json
… .github/scripts/data/networking-intelligence/test-validation-1754204905935.json
… .github/scripts/data/networking-intelligence/test-validation-1754205858038.json